Move section grabbing to store method
